Sensoriamento OK

+++ components/esp_nextion/include/esp_nextion.h
@@ -1,7 +1,9 @@
 #ifndef ESP_NEXTION_H
 #define ESP_NEXTION_H
 
-void nex_text_change_temp(char *buf, int tNumber, uint8_t value);
+#include <inttypes.h>
+
+void nex_text_change_temp(char *buf, int tNumber, int value);
 void nex_number_change_temp(char *buf, int nNumber, uint8_t value);
 void nex_pic_change_pic_id(char *buf, int pNumber, uint8_t value);
 void nex_button_change_pic_id(char *buf, int bNumber, uint8_t value);
